% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{audio_filter}
\alias{audio_filter}
\title{Apply high-pass filtering to extracted sound files}
\usage{
audio_filter(
  input_dir,
  output_dir = file.path(input_dir, "filtered"),
  hpf = 20 * 1000,
  normalise = FALSE
)
}
\arguments{
\item{input_dir}{directory containing wave files to filter}

\item{output_dir}{output directory. defaults to file.path(input_dir,"filtered")}

\item{hpf}{high-pass filtering frequency in Hz}

\item{normalise}{logical}
}
\description{
Filtering currently only implemented for mono wave files
}
