% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AudioMothSetup.R
\name{AudioMoth_days}
\alias{AudioMoth_days}
\title{Compute AudioMoth recording days}
\usage{
AudioMoth_days(mAH.d = 90, mAH = 2600, MB = 128 * 1000, MB.d = 3456)
}
\arguments{
\item{mAH.d}{Predicted daily energy consumption in mAH}

\item{mAH}{capacity of installed batteries}

\item{MB}{capacity of micro SD memory card in MB}

\item{MB.d}{daily file size in MB}
}
\value{
data frame
}
\description{
Approximates the possible recording duration in days, taking energy consumption and file size into account. Estimated file size and energy usage can be calculated using the `AudioMoth Configuration App` (see reference below)
}
\examples{
AudioMoth_days(mAH = 2900, mAH.d = 10, MB.d = 173)
}
\references{
https://github.com/OpenAcousticDevices/AudioMoth-Configuration-App
}
