% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BattyBirdNET.R
\name{BattyBirdNET}
\alias{BattyBirdNET}
\title{Processing data analysed using BattyBirdNET-Analyzer}
\usage{
BattyBirdNET(path = NULL, recursive = FALSE, meta = NULL, am_config = TRUE)
}
\arguments{
\item{path}{path to text files}

\item{recursive}{logical. Should the listing recurse into directories?}

\item{meta}{optional. data.frame with recording metadata}

\item{am_config}{logical. attempt to read Audiomoth configuration file if TRUE}
}
\description{
Processing data analysed using BattyBirdNET-Analyzer
}
\examples{
\dontrun{BirdNET(path = "Path to files")}
}
