% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET_tidyup.R
\name{BattyBirdNET_tidyup}
\alias{BattyBirdNET_tidyup}
\title{Performs a cleanup of sound archives}
\usage{
BattyBirdNET_tidyup(path, recursive)
}
\arguments{
\item{path}{path to text files}

\item{recursive}{logical. Should the listing recurse into directories?}
}
\description{
Performs a cleanup of sound archives
}
\details{
This function is called once files were analysed using BirdNET-Analyzer. Then empty BirdNET.results files are erased as well as wave files containing no audio
}
