% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET.R
\name{BirdNET}
\alias{BirdNET}
\title{Processing data analysed using BirdNET-Analyzer}
\usage{
BirdNET(
  path = NULL,
  recursive = FALSE,
  meta = NULL,
  am_config = FALSE,
  model = c("BirdNET v2.4", "Perch v2")
)
}
\arguments{
\item{path}{path to text files}

\item{recursive}{logical. Should the listing recurse into directories?}

\item{meta}{optional. data.frame with recording metadata}

\item{am_config}{logical. attempt to read Audiomoth configuration file if TRUE}

\item{model}{One of \code{c('BirdNET v2.4', 'Perch v2')}}
}
\description{
Processing data analysed using BirdNET-Analyzer
}
\examples{
\dontrun{BirdNET(path = "Path to files")}
}
