% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET_archive.R
\name{BirdNET_archive}
\alias{BirdNET_archive}
\title{Archive validated records}
\usage{
BirdNET_archive(
  BirdNET_results = NULL,
  path2archive = NULL,
  trash = TRUE,
  NocMig = TRUE,
  keep.false = FALSE,
  db = NULL
)
}
\arguments{
\item{BirdNET_results}{path to verified BirdNET.xlsx file}

\item{path2archive}{path to archive audio data}

\item{trash}{logical. Delete raw files? defaults to FALSE}

\item{NocMig}{logical}

\item{keep.false}{keep audio classified as false positive? defaults to FALSE}

\item{db}{local database (.xlsx) to which verified records are added}
}
\description{
Archive validated records
}
\details{
(1) Load table containing validated data
(2) Copy extracted events for future references to local database, distinguishing between true and false positives
(3) Providing NocMig summary if \code{NocMig = TRUE} as hourly totals per species
(4) Assembling string denoting hourly counts per species ready for submitting observation lists to ornitho
}
