% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET_archive_audiomoth.R
\name{BirdNET_archive_am}
\alias{BirdNET_archive_am}
\title{Archive validated records of AudioMoth recordings}
\usage{
BirdNET_archive_am(
  BirdNET_results = NULL,
  path2archive = NULL,
  keep.false = FALSE,
  db = NULL
)
}
\arguments{
\item{BirdNET_results}{path to verified BirdNET.xlsx file}

\item{path2archive}{path to archive audio data}

\item{keep.false}{keep audio classified as false positive? defaults to FALSE}

\item{db}{local database (.xlsx) to which verified records are added}
}
\description{
Archive validated records of AudioMoth recordings
}
\details{
(1) Load table containing validated data --> xlsx spreadsheet
(2) Copy extracted events for future reference to a local database, distinguishing between true and false positives (unverified records are ignored)
}
