% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET_cleanup.R
\name{BirdNET_cleanup}
\alias{BirdNET_cleanup}
\title{Cleanup folder: \strong{This will delete all files!}}
\usage{
BirdNET_cleanup(path = NULL, cleanup = FALSE)
}
\arguments{
\item{path}{path}

\item{cleanup}{logical. defaults to FALSE. Set to TRUE to actually remove files!}
}
\description{
Removes all files in a folder
}
