% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET_extract.R
\name{BirdNET_extract}
\alias{BirdNET_extract}
\title{Extract specific event from BirdNET results file}
\usage{
BirdNET_extract(
  path = NULL,
  taxon = NULL,
  score = NULL,
  nmax = NULL,
  sec = 1,
  output = NULL,
  spectro = FALSE,
  hyperlink = F
)
}
\arguments{
\item{path}{path. Relative path will be converted to absolute path.}

\item{taxon}{optional. select target taxa}

\item{score}{minimum confidence score}

\item{nmax}{maximum number of segments per taxon. defaults to NULL}

\item{sec}{seconds before and after target to extract}

\item{output}{path to write wav files}

\item{spectro}{logical. export spectro using \code{\link[bioacoustics]{spectro}} defaults to FALSE}

\item{hyperlink}{optional. Insert hyperlink to audio file in xlsx document. Only if records are not filtered.}
}
\description{
Extract events detected by BirdNET-Analyzer from original recording files. As labels species names and a time-stamp are used. As guidance for manual inspection of results, a maximum of 10 samples per taxon are randomly selected for validation. Selected segments are highlighted using the keyword 'validate' in the Quality column of the resulting `BirdNET.xlsx` file. Furthermore, the Comments column is populated with a consecutive numbering (1:N) per taxon.
}
\details{
There are a couple of important prerequisites:

(1) Run analyzer.py to detect events, making sure that --i and --o specify the same location.
(2) Reshape BirdNET results using function \code{\link{BirdNET}}
}
