% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET_figure.R
\name{BirdNET_figure}
\alias{BirdNET_figure}
\title{Visualise BirdNET results}
\usage{
BirdNET_figure(path, taxa)
}
\arguments{
\item{path}{path to BirdNET.xlsx file}

\item{taxa}{character}
}
\description{
Visualise BirdNET results
}
