% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET_hide.R
\name{BirdNET_hide}
\alias{BirdNET_hide}
\title{Hide wav files from BirdNET analyze.py}
\usage{
BirdNET_hide(path, hidden.path)
}
\arguments{
\item{path}{path}

\item{hidden.path}{path}
}
\description{
Temporarily move all wav file that where already processed, i.e. `BirdNET.results.txt` exists to another location to avoid rerunning analysis by `BirdNET_analyzer`.
Ideally, this is coded directly in analyze.py or its auxiliary functions defined in utils.py
}
