% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{BirdNET_labels2results}
\alias{BirdNET_labels2results}
\title{Read BirdNET_labels.txt}
\usage{
BirdNET_labels2results(path, recursive = FALSE)
}
\arguments{
\item{path}{path}

\item{recursive}{logical}
}
\description{
Read BirdNET_labels.txt
}
\keyword{internal}
