% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET.R
\name{BirdNET_man_detec}
\alias{BirdNET_man_detec}
\title{Append manually recovered detections to existing xlsx file}
\usage{
BirdNET_man_detec(path, recursive = FALSE)
}
\arguments{
\item{path}{path to text files}

\item{recursive}{logical}
}
\description{
Append manually recovered detections to existing xlsx file
}
\details{
Reads audacity labels from a text file and adds to BirdNET.xlsx
}
