% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET_meta.R
\name{BirdNET_meta}
\alias{BirdNET_meta}
\title{Write metadata of a recording project}
\usage{
BirdNET_meta(
  Location = NA,
  Lat = NA,
  Lon = NA,
  Device = NA,
  Micro = NA,
  Min_conf = 0.7,
  Overlap = 0,
  Sensitivity = 1.25,
  Slist = "BirdNET_V2.4"
)
}
\arguments{
\item{Location}{location name}

\item{Lat}{latitude}

\item{Lon}{longitude}

\item{Device}{recording device}

\item{Micro}{external microphone. default 'none'}

\item{Min_conf}{min_conf used in BirdNET-Analyzer}

\item{Overlap}{overlap used in BirdNET-Analyzer}

\item{Sensitivity}{sensitivity used in BirdNET-Analyzer}

\item{Slist}{path to slist used in BirdNET-Analyzer. default full model \code{'BirdNET_V2.4'}}
}
\value{
data frame
}
\description{
Add metadata to recording summary. Including
(1) Recording location and parameters
(2) BirdNET analysis settings
}
