% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET_quality.R
\name{BirdNET_quality}
\alias{BirdNET_quality}
\title{Try to approximate signal-to-noise (SNR) and loudness indices for BirdNET_results}
\usage{
BirdNET_quality(path, cores = parallel::detectCores() - 2, mar = 0.5)
}
\arguments{
\item{path}{folder containing wave files}

\item{cores}{number of cores}

\item{mar}{numeric vector of length 1. Specifies the margins adjacent to
the start and end points of selection over which to measure noise.}
}
\description{
(1) Computing SNR using \code{\link[warbleR]{sig2noise}}. As BirdNET gives detections within imprecise 3 sec intervals events are automatically narrowed down using \code{\link[warbleR]{auto_detec}}
(2) Computing loudness using \code{\link[soundgen]{getLoudness}}
}
