% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET.R
\name{BirdNET_results2txt}
\alias{BirdNET_results2txt}
\title{Modify 'BirdNET.results.txt' created by BirdNET-Analyzer}
\usage{
BirdNET_results2txt(
  path = NULL,
  recursive = FALSE,
  model = c("BirdNET v2.4", "Perch v2")
)
}
\arguments{
\item{path}{path to text files}

\item{recursive}{logical. Should the listing recurse into directories?}

\item{model}{One of \code{c('BirdNET v2.4', 'Perch v2')}}
}
\value{
data frame
}
\description{
Tweaks audacity labels by composing a string consisting of species name, recording date and time and detection score (e.g. "Blackbird 2023-08-15 [0.89]").
}
\details{
#' Reads text file containing audacity labels that were created by running the BirdNET analyzer script. Note, specifying rtype 'audacity' is required. Output is reformatted by adding the correct time stamp (estimated from file names) to the detected events. Empty files are ignored
}
\keyword{internal}
