% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET_snr.R
\name{BirdNET_snr}
\alias{BirdNET_snr}
\title{Try to approximate signal-to-noise (SNR) ratios for BirdNET_results}
\usage{
BirdNET_snr(path = NULL, mar = 0.5)
}
\arguments{
\item{path}{path. Relative path will be converted to absolute path.}

\item{mar}{numeric vector of length 1. Specifies the margins adjacent to
the start and end points of selection over which to measure noise.}
}
\description{
Computing SNR using \code{\link[warbleR]{sig2noise}}. As BirdNET gives detections within imprecise 3 sec intervals events are automatically narrowed down using \code{\link[warbleR]{auto_detec}}
}
