% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET_species.list.R
\name{BirdNET_species.list}
\alias{BirdNET_species.list}
\title{Create custom species list by picking species from `BirdNET_GLOBAL_6K_V2.4_Labels`}
\usage{
BirdNET_species.list(
  names = NULL,
  sciNames = FALSE,
  lang = "de",
  BirdNET_path = "../BirdNET-Analyzer/",
  species_list = "my_custom_species_list.txt",
  .write_text = TRUE
)
}
\arguments{
\item{names}{species names as character vector. Either scientific name or common names according to specific \code{lang}.}

\item{sciNames}{logical. default \code{FALSE}}

\item{lang}{Two-letter language code denoting a label file available for [BirdNET_Global_6K_V2.4](https://github.com/kahst/BirdNET-Analyzer/tree/main/labels/V2.4)}

\item{BirdNET_path}{path to BirdNET-Analyzer.}

\item{species_list}{path for exporting the species list}

\item{.write_text}{logical. defaults to \code{TRUE}}
}
\value{
data frame
}
\description{
Loads \strong{BirdNET_GLOBAL_6K_V2.4_Labels} from [BirdNET-Analyzer](https://github.com/kahst/BirdNET-Analyzer/tree/main/labels/V2.4) and filters for target species.
}
\details{
Requires that BirdNET_GLOBAL_6K_V2.4_Labels are locally available. The easiest way is to simply clone the entire [BirdNET-Analyzer] (https://github.com/kahst/BirdNET-Analyzer) repository. Alternatively manually download the needed folders [checkpoints/V2.4](https://github.com/kahst/BirdNET-Analyzer/tree/main/checkpoints/V2.4) and [labels/V2.4](https://github.com/kahst/BirdNET-Analyzer/tree/main/labels/V2.4) and save them within one directory, maintaining the original subfolders.
}
\examples{
\dontrun{BirdNET_species.list(names = c("Accipiter nisus"), sciNames = T, .write_text = F)}

}
