% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET.R
\name{BirdNET_table}
\alias{BirdNET_table}
\title{Data frame of BirdNET detections}
\usage{
BirdNET_table(
  path = NULL,
  recursive = FALSE,
  model = c("BirdNET v2.4", "Perch v2")
)
}
\arguments{
\item{path}{path}

\item{recursive}{logical. Should the listing recurse into directories?}

\item{model}{One of \code{c('BirdNET v2.4', 'Perch v2')}}
}
\description{
Data frame of BirdNET detections
}
\keyword{internal}
