% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET.R
\name{BirdNET_tidyup}
\alias{BirdNET_tidyup}
\title{Performs a cleanup of sound archives}
\usage{
BirdNET_tidyup(path, recursive, model = c("BirdNET v2.4", "Perch v2"))
}
\arguments{
\item{path}{path to text files}

\item{recursive}{logical. Should the listing recurse into directories?}

\item{model}{One of \code{c('BirdNET v2.4', 'Perch v2')}}
}
\description{
Performs a cleanup of sound archives
}
\details{
This function is called once files were analysed using BirdNET-Analyzer. Then empty BirdNET.results files are erased as well as wave files containing no audio
}
