% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET_hide.R
\name{BirdNET_unhide}
\alias{BirdNET_unhide}
\title{Unhide wav files from BirdNET analyze.py}
\usage{
BirdNET_unhide(path, hidden.path)
}
\arguments{
\item{path}{path}

\item{hidden.path}{path}
}
\description{
Unhide wav files from BirdNET analyze.py
}
