% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_wav_header}
\alias{check_wav_header}
\title{Check for invalid RIFF file}
\usage{
check_wav_header(file)
}
\arguments{
\item{file}{wave file}
}
\description{
Check for invalid RIFF file
}
\keyword{internal}
