% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chirpity2audacity.R
\name{chirpity2audacity}
\alias{chirpity2audacity}
\title{Convert chirpity results to audacity labels}
\usage{
chirpity2audacity(path)
}
\arguments{
\item{path}{path to chirpity results file (.csv)}
}
\description{
Convert chirpity results to audacity labels
}
