% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chirpity2BirdNET.R
\name{chirpity2BirdNET}
\alias{chirpity2BirdNET}
\title{Combine event detections of [BirdNET](github.com/kahst/BirdNET-Analyzer) and [chirpity](https://chirpity.mattkirkland.co.uk/)}
\usage{
chirpity2BirdNET(
  path = NULL,
  birdnet = file.path(path, "BirdNET.xlsx"),
  chirpity = file.path(path, "Chirpity.csv"),
  output = file.path(path, "BirdNET_Chirpity.xlsx"),
  ignore_terms = c("call", "flight call", "song")
)
}
\arguments{
\item{path}{path. Relative path will be converted to absolute path.}

\item{birdnet}{path to BirdNET.xlsx file}

\item{chirpity}{path to chirpity.csv file}

\item{output}{defaults to file.path(path, "BirdNET_Chirpity.xlsx")}

\item{ignore_terms}{character vector of text to remove when in parentheses (e.g. '(call)' or '(flight call)')}
}
\description{
Merges detection tables created by [BirdNET](github.com/kahst/BirdNET-Analyzer) and [chirpity](https://chirpity.mattkirkland.co.uk/) models. Events are classified as identical when Taxon and Event timestamp T1 are the same. Because [chirpity](https://chirpity.mattkirkland.co.uk/) further differentiates certain call types, extensions such as 'call' or 'flight call' are ignored when specified by param ignore_terms
}
