% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chirpity_extract.R
\name{chirpity_extract}
\alias{chirpity_extract}
\title{Extract events detected by [chirpity](https://chirpity.mattkirkland.co.uk/) and append to `BirdNET.xlsx`}
\usage{
chirpity_extract(
  path = NULL,
  input = file.path(path, "BirdNET_Chirpity.xlsx"),
  output = NULL,
  spectro = FALSE,
  hyperlink = FALSE,
  sec = 1
)
}
\arguments{
\item{path}{path. Relative path will be converted to absolute path.}

\item{input}{xlsx file containing BirdNET and chirpity results}

\item{output}{path to write wav files}

\item{spectro}{logical. export spectro using \code{\link[bioacoustics]{spectro}} defaults to FALSE}

\item{hyperlink}{optional. Insert hyperlink to audio file in xlsx document. Only if records are not filtered.}

\item{sec}{seconds before and after target to extract}
}
\description{
see \link{BirdNET_extract}
}
\details{
extract events exclusively detected by [chirpity](https://chirpity.mattkirkland.co.uk/) and add to `BirdNET.xlsx`
}
