% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dusk2dawn.R
\name{dusk2dawn}
\alias{dusk2dawn}
\title{Compose dusk-to-dawn string based on date and location}
\usage{
dusk2dawn(date = Sys.Date(), lat = 52.032, lon = 8.517, tz = "CET")
}
\arguments{
\item{date}{Date. Default \code{\link[base]{Sys.Date}}}

\item{lat}{latitude in decimal degrees}

\item{lon}{longitude in decimal degrees}

\item{tz}{timezone. Default CEt}
}
\description{
Calling \code{\link[suncalc]{getSunlightTimes}} to retrieve times of dusk and dawn for a coordinate pair
}
\examples{
## usage
dusk2dawn(date = Sys.Date(), lat = 52, lon = 8)
}
