% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NocMig_process.R
\name{NocMig_process}
\alias{NocMig_process}
\title{Process NocMig session for downstream analysis}
\usage{
NocMig_process(
  parent.folder = "E:/NocMig",
  child.folder = "",
  format = c("wav", "mp3"),
  rename = TRUE,
  time_reference = c("first", "each"),
  ctime = TRUE,
  lat = 52.03209,
  lon = 8.516775,
  segment_length = NULL
)
}
\arguments{
\item{parent.folder}{path}

\item{child.folder}{optional. defaults to an emoty string \code{''}}

\item{format}{audio format. Supported are \code{'wav'} and \code{'mp3'}}

\item{rename}{logical. defaults to \code{TRUE}}

\item{time_reference}{(1) Specify \code{'first'} when all recordings share ctime/mtime of the first recording (e.g. \strong{Olympus LS recorder}). (2) Specify \code{'each'} when recordings have unique ctime/mtime (e.g. \strong{Sony PCM recorder})}

\item{ctime}{logical. defaults to \code{TRUE}. If \code{FALSE} mtime is used instead.}

\item{lat}{latitude in decimal degrees}

\item{lon}{longitude in decimal degrees}

\item{segment_length}{optional numeric value. If specified, function \code{\link{split_wave}} is used to to segmenti long audio files in shorter segments. defaults to \code{NULL}}
}
\description{
Process NocMig session for downstream analysis
}
