% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read.audacity}
\alias{read.audacity}
\title{Replacement for \code{\link[seewave]{read.audacity}} to avoid problems with certain labels containing apostrophes etc.}
\usage{
read.audacity(
  file,
  delim = "\\t",
  col_names = c("start", "end", "label", "score")
)
}
\arguments{
\item{file}{A .txt file}

\item{delim}{delim}

\item{col_names}{col_names}
}
\description{
Read content of a text file and export in format matching \code{\link[seewave]{read.audacity}}
}
