% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reformat_xlsx}
\alias{reformat_xlsx}
\title{Format BirdNET.xlsx in a nice style}
\usage{
reformat_xlsx(path, model = c("BirdNET v2.4", "Perch v2"))
}
\arguments{
\item{path}{path}

\item{model}{One of \code{c('BirdNET v2.4', 'Perch v2')}}
}
\description{
Format BirdNET.xlsx in a nice style
}
