% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_decimal_seconds}
\alias{remove_decimal_seconds}
\title{Eliminate postfix decimal seconds from file names}
\usage{
remove_decimal_seconds(wave_file, input_dir)
}
\arguments{
\item{wave_file}{file name}

\item{input_dir}{path to file to check file extension}
}
\description{
Eliminate postfix decimal seconds from file names
}
\keyword{internal}
