% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename2DateTime}
\alias{rename2DateTime}
\title{Rename file names to YYYYMMDD_HHMMSS format
remove junk from file names and keep only date & time (ie last 15 characters YYYYMMDD_HHMMSS)
do not change files in subfolder extracted}
\usage{
rename2DateTime(input_dir, pattern = ".WAV")
}
\arguments{
\item{input_dir}{input folder}

\item{pattern}{file extention to look for}
}
\description{
Rename file names to YYYYMMDD_HHMMSS format
remove junk from file names and keep only date & time (ie last 15 characters YYYYMMDD_HHMMSS)
do not change files in subfolder extracted
}
