% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{total_duration}
\alias{total_duration}
\title{Compute entire recording length}
\usage{
total_duration(path, format = "WAV", recursive = FALSE)
}
\arguments{
\item{path}{Path to folder containing audio files}

\item{format}{audio format. Supported are \code{'wav'} and \code{'mp3'}}

\item{recursive}{logical. Should the listing recurse into directories?}
}
\description{
Compute entire recording length
}
\keyword{internal}
