% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xlsx_append}
\alias{xlsx_append}
\title{Append columns to existing xslx database}
\usage{
xlsx_append(path, data, model = c("BirdNET v2.4", "Perch v2"))
}
\arguments{
\item{path}{file}

\item{data}{data frame}

\item{model}{One of \code{c('BirdNET v2.4', 'Perch v2')}}
}
\description{
Append columns to existing xslx database
}
