% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET_man_detec.R
\name{BirdNET_man_detec}
\alias{BirdNET_man_detec}
\title{Append manually recovered detections to existing xlsx file}
\usage{
BirdNET_man_detec(path, spectro = FALSE)
}
\arguments{
\item{path}{path to text files}

\item{spectro}{logical. export spectro using \code{\link[bioacoustics]{spectro}} defaults to FALSE}
}
\description{
Append manually recovered detections to existing xlsx file
}
