% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{BirdNET_name_repair}
\alias{BirdNET_name_repair}
\title{Cleanup taxon names to avoid problems when specifying path}
\usage{
BirdNET_name_repair(path)
}
\arguments{
\item{path}{path}
}
\description{
Fix the slash '/' in the taxon name and replace by a minus '-'-
}
\keyword{internal}
