% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{BirdNET_table}
\alias{BirdNET_table}
\title{Data frame of BirdNET detections}
\usage{
BirdNET_table(path = NULL, recursive = FALSE)
}
\arguments{
\item{path}{path}

\item{recursive}{logical. Should the listing recurse into directories?}
}
\description{
Data frame of BirdNET detections
}
\keyword{internal}
