% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_timestamp}
\alias{get_timestamp}
\title{Extract datetime from file name}
\usage{
get_timestamp(x)
}
\arguments{
\item{x}{file name}
}
\value{
datetime object
}
\description{
Extract date and time from file name, assuming a prefix of the form `YYYYMMDD_HHMMSS`
}
