% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NocMig_process.R
\name{NocMig_process}
\alias{NocMig_process}
\title{Process NocMig session for downstream analysis}
\usage{
NocMig_process(
  path = "E:/NocMig",
  folder = NULL,
  format = c("WAV", "wav", "mp3", "MP3"),
  ctime = c("first", "each"),
  lat = 52.03209,
  lon = 8.516775
)
}
\arguments{
\item{path}{path to NocMig folders}

\item{folder}{folder to process}

\item{format}{audio format. Supported are \code{'wav'} and \code{'mp3'}}

\item{ctime}{(1) Specify \code{'first'} when all recordings share ctime of the first recording (e.g. \strong{Olympus LS recorder}). (2) Specify \code{'each'} when recordings have unique ctime (e.g. \strong{Sony PCM recorder})}

\item{lat}{latitude in decimal degrees}

\item{lon}{longitude in decimal degrees}
}
\description{
Process NocMig session for downstream analysis
}
